requirements.txt
beautifulsoup4>=4.12
playwright>=1.40
pytest>=8.0
python-dotenv>=1.0
python-telegram-bot>=21
requests>=2.31
urllib3>=2.0
